Add box2vmid table to support existing molecule files
We already have a set of molecule files created for vagrant, which have
the 'box' parameter to specify the template name.  Since it's not
possible to override these in the base config, as a workaround, add a
lookup table in the driver options which can be used to lookup a
template_id from a box name.