Set instance hostnames
Use the hostname module to set the instance hostname after new instances
are cloned.  It seems this is not easy to do with the proxmox cloudinit,
so just use the Ansible hostname module in the prepare playbook to set
the hostnames of the new instances.

This adds some more time to the creation process, so if you don't care
about the hostname on the molecule instance, set the 'sethostname'
driver option to false to disable this for all instances, or
'sethostname' to false on individual instances in the platforms section.
Also, you can set 'hostname' in the platform to use a different hostname
than the molecule instance name.

This feature is omitted if you provide a custom prepare.yml playbook in
your molecule scenario, since that overrides the driver provided
playbook. If you want to set the hostname in that case, you'll need to
add a hostname task to your custom scenario playbook.