Fix set hostname task in prepare playbook
The prepare playbook runs like a normal playbook over the inventory, not
like the create and destroy playbooks which to loop over the molecule
platforms.

Remove the incorrect loop in the hostname task in the playbook. This
only works when there is exactly one molecule instance in the playbook.

Set the hostname to the inventory_hostname, which is the same name
as the molecule platform instance name.

If you want to override the target hostname, add the 'proxmox_hostname'
as a host variable in your inventory.