Clone by template name
We need to update the proxmox templates from time to time. Since
templates are read-only, the only way to do that is to clone the
original template, make our changes in the cloned machine, then convert
the machine into template.  We can rename templates, but cannot renumber
them.  The upshot is, it is better to specify the templates by name
instead of vmid if we want to avoid churn in the molecule.yml file.

To support this, add the new 'proxmox_template_name' keyword to the
platform instances to specify the name of the template to be cloned.
(Also, support the shorter name 'template_name' as an alias.). The
default template name is 'molecule', and that can be overrided by
setting the 'template_name' in the driver options.

Continue to support cloning by vmid, and that takes precedence if both
the name and the vmid is given.

Drop support of the 'box2vmid' option, which was a work-around some
existing vagrant-based molecule.yml files I had.