Fix missing vm start after clone
Change our custom proxmox_qemu_agent module to start the vm when it
detects the vm is not running.  Reorganize the qemu guest agent query
retry code to be more clear.

Remove the work-around delay and the proxmox_kvm 'start' task in the
create playbook.  This was found to be unreliable when recycling vmids.
It seems the vm status can have the old 'running' status for a small
amount of time when vmid's are recycled quickly.

Note: Unfortunately, the proxmox_kvm module tries to start the template
(which fails, as expected) when the state parameter is 'started',
instead of cloning then starting the cloned vm, so we need to start the
vm after the proxmox_kvm module. If this changes in the future, we can
update the state parameter so proxmox_kvm starts the clone. No changes
to the proxmox_qemu_agent should be needed.

Fixes #1