Use platform names to populate the instance_config
The value of the name field in the instance_config file must match the
names in the platforms in the molecule.yml file.  After removing the
long workaround delay after the clone, sometimes the instance_config
names are set to "VM <number>" instead of the correct name, and so ssh
accesses fail during the prepare, converge, and verify plays.

Currently, the vm name is set when the guest is cloned, then
the status of the vm queried in the next task. The result of that
query is used to populate the name field in the instance_config.
Unfortunately, the status may not contain the name yet, depending on
timing.

Instead, propagate the platform name from the molecule.yml file to the
instance_config using the registered task variables.  It is a bit
tricky, since the names are in the results of a result, so use short and
unique names for the loop variables of each task, and avoid using 'vm'
as a loop variable.

Also, use the loop label option to reduce the noisy output verbosity in
the create and destroy playbooks,