tests: Run pytest in virtualenv
It is not enough to run the pytest file in the virtualenv, the
virtualenv must be activated first, so we run the molecule installed in
our venv, since it is run as a separate process, not a python module.

Update the makefile test and check targets to be sure the venv is active
before running pytest.