Also update the /etc/hosts when setting the hostname
Remove the workaround loopback 127.0.1.1 address if present in the
/etc/hosts file. This is present in debian based systems and can
interfere with some tests.

Be sure to add our address/hostname to the /etc/hosts since we usually
do not have the ephemeral virtual machine in DNS.  Run setup to gather
the machines IPv4 address to put in the /etc/hosts file.