Use the vmid to set cloud-init values
The cloud-init values are set after the initial clone, but
the instance name may not be ready yet, so use the vmid
returned from the clone results when setting the cloud-init
values.

Flatten the platform settings to make it easier to use the
new cloud-init keys (and make this more like other drivers).

Add a cloud-init unit test.

Mention a cloud-init disk is required in the template if you
want to use any of the cloud-init settings.