Rename driver to molecule-proxmox
Rename this driver from 'proxmox' to 'molecule-proxmox' so molecule will
skip the built-in schema verification and allow the driver to run.

This allows the proxmox driver to be run by a molecule version newer
than 4.0.1, which was the last version that worked before the molecule
project changed how schema validation works and broke all of the
non-core drivers.  Later, support for "custom drivers" was added in the
molecule project by treating drivers that start with 'molecule-' or
'custom-' to be considerd "custom drivers" and the build-in schema
checks are skipped.

This change means you will need to update your molecule.yml file when
upgrading from 4.0.1 to a more recent version of molecule. New molecule
files should use the 'molecule-proxmox' driver name.

For now, molecule will print a warning message, since the schema
validation is skipped. Later we should add driver specific schema
validation.