Move virtual machine creation tasks to task file
Move the tasks to create and start kvm instances to a task file and
include the task once for each item in the molecule platforms list,
instead of running each task in a loop, and passing the nested results
to each once in order to populate the molecule instance_config yaml
file.

This should make it easier to debug and maintain the create playbook.