Bump molecule version
Remove the molecule version cap from setup.

Add molecule 24.2.0 and 24.2.1 to the tox tests.

Add a comment in the tox config to mention we need to install the docker
plugin to work-around a molecule bug.  The issue is docker is (was?) the
default plugin so is unfortunately required for molecule reset.