Add requests to setup.py
Proxmoxer supports different authentication backends and therefore does not require requests to be installed. But the proxmox_kvm module only supports http-requests, while not being and pip module and not being able to install a package.

Since molecule-proxmox requires proxmox_kvm, it should require requests as well, to work out of the box after installing it into an empty virtualenv.