Run proxmox_secrets when it is executable
When the proxmox_secrets option specifies an executable file, run that
file and capture the stdout as yaml to be used for driver settings.

This makes it possible use external password managers such as `pass` and
`gpg` to store proxmox api passwords and connection tokens.  Normally, a
script should be provided to run `pass`, `gpg`, or some other secrets
management program with the correct arguments. The stdout should be in
yaml format to specify one or more driver options.

While here, add a better error messages if the proxmox_secret file is
not found. This should help debug issues when the file is missing or
just the wrong name or path is specified.

Add var-naming[no-jinja] to the ansible-lint skip_list so we can
dynamically set the api connection variables.