Add missing molecule versions to tox.ini
Synchronize the molecule versions in the tox.ini with the supported
version range specified in the setup.py file.

Add the python versions to test for each molecule version. Currently we
test with python 3.10 to 3.12 for older molecule versions, and python
3.11 to 3.13 for recent molecule versions.

The ansible version required depends on the molecule version, so make
the ansible version conditional as well.

While we are here, fix some typos in the comments.