Remove the .venv setup from the Makefile
We are not very sensitive to the tox version and these days it is easy
to install it with pipx, so just remove the virtualenv setup from the
Makefile to make it simpler.

It is normally impractical to run every testenv, so change `make test`
helper target to run just the "latest" testenv by default, and support
running other ones with the TESTENV variable. (All the testenv's can be
run by running tox directly if needed.)

Update the developer instructions in the README to explain that tox
needs to be installed, and explain how to create the test environment
variables by copying our sample envrc file.