tests: Optionally use linked clones in tests
Use the TEST_PROXMOX_FULL_CLONE env var to specify whether a full clone
should be created during the tests. The default is true (create a full
clone).  Set TEST_PROXMOX_FULL_CLONE to false to create linked clones
which can reduce the test times, depending on how fast the drives are on
the proxmox system.

Change the linked-clone scenario to always create a linked clone,
because, well, it's a linked-clone test.