Set the driver module path in ANSIBLE_LIBRARY
As of molecule 25.2.0, molecule no longer sets the ANSIBLE_LIBRARY
environment variable to include the module path declared by plugins by
the module_dir() method.

This was done as part of the effort to improve test-isolation in the
ansible tool suite.  See molecule commit 30a97494ebe6 (Remove code that
modifies ansible import paths (#4380))

Until we are able to remove our custom module, just set the module
search path by setting ANSIBLE_LIBRARY when the driver is loaded. This
allows the driver to continue to work with recent versions of molecule.