Add a minimal pyproject.toml file
Add a pyproject.toml file in the project root directory, since this is
expected by modern python packaging tools.  Add a minimal configuration
as shown in the setuptools documentation.

Keep setup.py for now, since setuptools still supports it as a first
class configuration file, but remove the dynamically generated
configuration from setup.py, except for the code to read the README.rst.
The long desription in used to create the PyPI project page, and so we
use the same test as the readme.

Add the `tox -e build` developer task to build the package using the
modern python -m build method.

Add a description setting to the `tox -e release` task.